spark支持变量逻辑

+++ spark-yun-backend/spark-yun-modules/src/main/java/com/isxcode/spark/modules/work/run/impl/SparkSqlExecutor.java
@@ -47,6 +47,7 @@ import com.isxcode.spark.modules.work.repository.WorkConfigRepository;
 import com.isxcode.spark.modules.work.repository.WorkInstanceRepository;
 import com.isxcode.spark.modules.work.repository.WorkRepository;
 import com.isxcode.spark.modules.work.run.WorkExecutor;
+import com.isxcode.spark.modules.work.run.WorkInfo;
 import com.isxcode.spark.modules.work.run.WorkRunContext;
 import com.isxcode.spark.modules.work.sql.SqlCommentService;
 import com.isxcode.spark.modules.work.sql.SqlFunctionService;
@@ -114,13 +115,14 @@ public class SparkSqlExecutor extends WorkExecutor {
     private final SecretKeyRepository secretKeyRepository;
 
     public SparkSqlExecutor(WorkInstanceRepository workInstanceRepository, ClusterRepository clusterRepository,
-                            ClusterNodeRepository clusterNodeRepository, WorkflowInstanceRepository workflowInstanceRepository,
-                            WorkRepository workRepository, WorkConfigRepository workConfigRepository, Locker locker,
-                            HttpUrlUtils httpUrlUtils, FuncRepository funcRepository, FuncMapper funcMapper,
-                            ClusterNodeMapper clusterNodeMapper, AesUtils aesUtils, IsxAppProperties isxAppProperties,
-                            FileRepository fileRepository, DatasourceService datasourceService, SqlCommentService sqlCommentService,
-                            SqlValueService sqlValueService, SqlFunctionService sqlFunctionService, AlarmService alarmService,
-                            DatasourceMapper datasourceMapper, DataSourceFactory dataSourceFactory, SecretKeyRepository secretKeyRepository) {
+        ClusterNodeRepository clusterNodeRepository, WorkflowInstanceRepository workflowInstanceRepository,
+        WorkRepository workRepository, WorkConfigRepository workConfigRepository, Locker locker,
+        HttpUrlUtils httpUrlUtils, FuncRepository funcRepository, FuncMapper funcMapper,
+        ClusterNodeMapper clusterNodeMapper, AesUtils aesUtils, IsxAppProperties isxAppProperties,
+        FileRepository fileRepository, DatasourceService datasourceService, SqlCommentService sqlCommentService,
+        SqlValueService sqlValueService, SqlFunctionService sqlFunctionService, AlarmService alarmService,
+        DatasourceMapper datasourceMapper, DataSourceFactory dataSourceFactory,
+        SecretKeyRepository secretKeyRepository) {
 
         super(workInstanceRepository, workflowInstanceRepository, alarmService, sqlFunctionService);
         this.workInstanceRepository = workInstanceRepository;
@@ -204,25 +206,36 @@ public class SparkSqlExecutor extends WorkExecutor {
             .mainClass("com.isxcode.spark.plugin.query.sql.Execute").appResource("spark-query-sql-plugin.jar")
             .conf(genSparkSubmitConfig(workRunContext.getClusterConfig().getSparkConfig())).build();
 
-        // 去掉sql中的注释
-        String sqlNoComment = sqlCommentService.removeSqlComment(workRunContext.getScript());
+        // 判断实例中是否有脚本
+        String script;
+        String printSql;
+        if (workInstance.getWorkInfo() == null) {
+            // 去掉sql中的注释
+            String sqlNoComment = sqlCommentService.removeSqlComment(workRunContext.getScript());
 
-        // 解析上游参数
-        String jsonPathSql = parseJsonPath(sqlNoComment, workInstance);
+            // 解析上游参数
+            String jsonPathSql = parseJsonPath(sqlNoComment, workInstance);
 
-        // 翻译sql中的系统变量
-        String parseValueSql = sqlValueService.parseSqlValue(jsonPathSql);
+            // 翻译sql中的系统变量
+            String parseValueSql = sqlValueService.parseSqlValue(jsonPathSql);
 
-        // 翻译sql中的系统函数
-        String script = sqlFunctionService.parseSqlFunction(parseValueSql);
-        String printSql = script;
+            // 翻译sql中的系统函数
+            script = sqlFunctionService.parseSqlFunction(parseValueSql);
+            printSql = script;
 
-        // 翻译全局变量
-        List<SecretKeyEntity> allKey = secretKeyRepository.findAll();
-        for (SecretKeyEntity secretKeyEntity : allKey) {
-            script = script.replace("${{ secret." + secretKeyEntity.getKeyName() + " }}",
-                secretKeyEntity.getSecretValue());
-            printSql = printSql.replace("${{ secret." + secretKeyEntity.getKeyName() + " }}", "******");
+            // 翻译全局变量
+            List<SecretKeyEntity> allKey = secretKeyRepository.findAll();
+            for (SecretKeyEntity secretKeyEntity : allKey) {
+                script = script.replace("${{ secret." + secretKeyEntity.getKeyName() + " }}",
+                    secretKeyEntity.getSecretValue());
+                printSql = printSql.replace("${{ secret." + secretKeyEntity.getKeyName() + " }}", "******");
+            }
+
+            workInstance.setWorkInfo(JSON.toJSONString(WorkInfo.builder().sparkSql(script).printSql(printSql).build()));
+        } else {
+            WorkInfo workInfo = JSON.parseObject(workInstance.getWorkInfo(), WorkInfo.class);
+            script = workInfo.getSparkSql();
+            printSql = workInfo.getPrintSql();
         }
 
         // 打印sql日志
